Remove obsolete Index and TagIndex; consolidate into NoteIndex
